Another fix from inspection

+++ source/blender/render/intern/texture_image.cc
@@ -698,7 +698,7 @@ static int ibuf_get_color_clip(float col[4], ImBuf *ibuf, int x, int y, int extf
   }
 
   if (ibuf->float_buffer.data) {
-    const float *fp = ibuf->float_buffer.data + (x + y * ibuf->x) * ibuf->channels;
+    const float *fp = ibuf->float_buffer.data + (x + size_t(y) * ibuf->x) * ibuf->channels;
     if (ibuf->channels == 1) {
       col[0] = col[1] = col[2] = col[3] = *fp;
     }
@@ -710,7 +710,7 @@ static int ibuf_get_color_clip(float col[4], ImBuf *ibuf, int x, int y, int extf
     }
   }
   else {
-    const uchar *rect = ibuf->byte_buffer.data + 4 * (x + y * ibuf->x);
+    const uchar *rect = ibuf->byte_buffer.data + 4 * (x + size_t(y) * ibuf->x);
     float inv_alpha_fac = (1.0f / 255.0f) * rect[3] * (1.0f / 255.0f);
     col[0] = rect[0] * inv_alpha_fac;
     col[1] = rect[1] * inv_alpha_fac;
